+++ CloudShell.ControlPlane.Client/ControlPlaneClientServiceCollectionExtensions.cs
@@ -134,6 +134,8 @@ public static class ControlPlaneClientServiceCollectionExtensions
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
         services.AddScoped<IMetricManager>(
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
+        services.AddScoped<IResourceHealthManager>(
+            serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
         services.AddScoped<IResourceMonitoringManager>(
             serviceProvider => serviceProvider.GetRequiredService<IControlPlane>());
     }