+++ CloudShell.Hosting/Hosting/CloudShellHostApplicationBuilderExtensions.cs
@@ -73,7 +73,6 @@ public static class CloudShellHostApplicationBuilderExtensions
 
         builder.Services.AddSingleton<ShellCatalog>();
         builder.Services.AddScoped<ICloudShellNavigator, CloudShellNavigator>();
-        builder.Services.AddScoped<ResourceHealthCheckService>();
         builder.Services.AddScoped<ResourceManagerDisplaySettings>();
     }